Move fuzzy_cli logic to a lib module
